backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
pandas==2.1.3
openpyxl==3.1.2
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
APScheduler==3.10.4
email-validator==2.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
databases[postgresql]==0.8.0
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1
email-validator==2.1.0
python-dateutil==2.8.2
cryptography==41.0.8
aiofiles==23.2.1
Pillow==10.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2